Refactor: event 객체 구조 리팩토링
