+++ src/js/index.js
@@ -1,4 +1,4 @@
-import { RacingEvent } from './Controller/Event/event.js';
+import { racingCarInputEvent } from './Controller/InputEvent/racingCarInputEvent.js';
 import {selector} from './util/consts.js'
 import Form from './View/Component/Form/Form.js';
 import { render } from './View/index.js';
@@ -12,8 +12,13 @@ const carForm = selector('.racing-form')
 carForm.addEventListener('submit', event => event.preventDefault())
 
 const carNameButton = selector('.car-name-button');
-carNameButton.addEventListener('click', RacingEvent.carNameClickEvent)
+carNameButton.addEventListener('click', racingCarInputEvent.racingCarNameClickEvent)
 
 const carNameInput = selector('.car-name-input');
-carNameInput.addEventListener('keypress', RacingEvent.carNameKeyboardEvent)
+carNameInput.addEventListener('keypress', racingCarInputEvent.racingCarNameKeyboardEvent)
 
+const racingCountButton = selector('.racing-count-button');
+// racingCountButton.addEventListener('click', )
+
+const racingCountInput = selector('.racing-count-input')
+// racingCountInput.addEventListener('keypress', )
